+++ openstudiocore/src/osversion/VersionTranslator.cpp
@@ -92,7 +92,7 @@ VersionTranslator::VersionTranslator()
   m_updateMethods[VersionString("1.2.3")] = &VersionTranslator::update_1_2_2_to_1_2_3;
   m_updateMethods[VersionString("1.3.5")] = &VersionTranslator::update_1_3_4_to_1_3_5;
   m_updateMethods[VersionString("1.5.4")] = &VersionTranslator::update_1_5_3_to_1_5_4;
-  m_updateMethods[VersionString("1.7.2")] = &VersionTranslator::defaultUpdate;
+  m_updateMethods[VersionString("1.7.2")] = &VersionTranslator::update_1_7_1_to_1_7_2;
 
   // List of previous versions that may be updated to this one.
   //   - To increment the translator, add an entry for the version just released (branched for
@@ -2380,6 +2380,46 @@ std::string VersionTranslator::update_1_5_3_to_1_5_4(const IdfFile& idf_1_5_3, c
     } else {
       ss << object;
 
+    }
+  }
+
+  return ss.str();
+}
+
+std::string VersionTranslator::update_1_7_1_to_1_7_2(const IdfFile& idf_1_7_1, const IddFileAndFactoryWrapper& idd_1_7_2)
+{
+  std::stringstream ss;
+
+  ss << idf_1_7_1.header() << std::endl << std::endl;
+
+  // new version object
+  IdfFile targetIdf(idd_1_7_2.iddFile());
+  ss << targetIdf.versionObject().get();
+
+  for (const IdfObject& object : idf_1_7_1.objects()) {
+    if (object.iddObject().name() == "OS:EvaporativeCooler:Direct:ResearchSpecial") {
+      auto newObject = object.clone(true);
+      auto d = object.getDouble(4);
+      if( ! d ) {
+        newObject.setString(4,"Autosize");
+      }
+      newObject.setDouble(11,0.1);
+
+      m_refactored.push_back( std::pair<IdfObject,IdfObject>(object,newObject) );
+      ss << newObject;
+    } else if (object.iddObject().name() == "OS:EvaporativeCooler:Indirect:ResearchSpecial") {
+      auto newObject = object.clone(true);
+      auto d = object.getDouble(5);
+      if( ! d ) {
+        newObject.setString(4,"Autosize");
+      }
+      newObject.setDouble(22,0.1);
+      newObject.setDouble(24,1.0);
+
+      m_refactored.push_back( std::pair<IdfObject,IdfObject>(object,newObject) );
+      ss << newObject;
+    } else {
+      ss << object;
     }
   }